+++ src/scorer/feature/chaos_scorer.ts
@@ -2,6 +2,7 @@
  * Random number generator to mix up the feed.
  */
 import FeatureScorer from '../feature_scorer';
+import Toot from '../../api/objects/toot';
 import { WeightName } from "../../types";
 
 
@@ -10,7 +11,24 @@ export default class ChaosScorer extends FeatureScorer {
         super({scoreName: WeightName.CHAOS});
     }
 
-    async _score() {
-        return Math.random();
+    async _score(toot: Toot) {
+        try {
+            return this.decimalHash(toot.content);
+        } catch (e) {
+            console.warn(`Error in _score() for ${this.name}:`, e);
+            console.warn(`Toot with error in ChaosScorer:`, toot);
+            return 0;
+        }
+    }
+
+    // Use a hash to get a deterministic score between 0 and 1
+    decimalHash(s: string): number {
+        let hash = 0;
+
+        for (let i = 0; i < s.length; i++) {
+            hash = s.charCodeAt(i) + ((hash << 5) - hash);
+        }
+
+        return (hash & hash) / Math.pow(2, 31);
     }
 };